feat: テストファイルを別フォルダに移動してレッド状態を追加 #64
