+++ test_basics/15_TDD/tdd-practice/tsconfig.json
@@ -10,5 +10,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules", "dist"]
 }
